tighten up memory allocations in places where they were unnecessarily large; add cudaDeviceSynchronize in some spots to catch kernel launch errors; improve screen formatting, comments and debugging output

+++ AmpTools/GPUManager/GPUManager.h
@@ -100,12 +100,12 @@ public:
                       size_t startEvent, size_t nEvents );
   
   // General utils:
-  static int calcNEventsGPU( int iNEvents ){
+  static size_t calcNEventsGPU( size_t iNEvents ){
     
     //Should be a power of 2 for reduction to work, also multiple of GPU_BLOCK_SIZE_SQ    
     int iPow = 0;
-    while( ( 1 << iPow ) < iNEvents ) iPow++;
-    return(  (1<<iPow) < GPU_BLOCK_SIZE_SQ ? GPU_BLOCK_SIZE_SQ : (1<<iPow) );
+    while( ( 1L << iPow ) < iNEvents ) iPow++;
+    return(  ( 1L << iPow ) < GPU_BLOCK_SIZE_SQ ? GPU_BLOCK_SIZE_SQ : ( 1L << iPow ) );
   }
   
   bool m_ownsData;
